update test onnx utils
